requirements.txt
Django==2.1.2
google-cloud-language==1.0.2
